Code cleanup
Removed unnecessary extra main file
Added conditional statement to end calculation to determine if the weighted values
should be considered or not
Cleaned up some variable names
